move NumpyEncoder to utils
