+++ bluecellulab/tools.py
@@ -15,7 +15,6 @@
 
 
 from __future__ import annotations
-import json
 import math
 import multiprocessing
 import multiprocessing.pool
@@ -635,20 +634,6 @@ def search_hyp_current_replay_imap(blueconfig, gid_list, timeout=600,
             pool.terminate()
             yield (None, None)
     pool.terminate()
-
-
-class NumpyEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, (np.int_, np.intc, np.intp, np.int8,
-                            np.int16, np.int32, np.int64, np.uint8,
-                            np.uint16, np.uint32, np.uint64)):
-            return int(obj)
-        elif isinstance(obj, (np.float_, np.float16, np.float32,
-                              np.float64)):
-            return float(obj)
-        elif isinstance(obj, np.ndarray):
-            return obj.tolist()
-        return json.JSONEncoder.default(self, obj)
 
 
 def check_empty_topology() -> bool:
