+++ bluecellulab/utils.py
@@ -1,6 +1,9 @@
 """Utility functions."""
 import contextlib
 import io
+import json
+
+import numpy as np
 
 
 def run_once(func):
@@ -23,3 +26,17 @@ class CaptureOutput(list):
     def __exit__(self, exc_type, exc_val, exc_tb):
         self._redirect_stdout.__exit__(exc_type, exc_val, exc_tb)
         self.extend(self._stringio.getvalue().splitlines())
+
+
+class NumpyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, (np.int_, np.intc, np.intp, np.int8,
+                            np.int16, np.int32, np.int64, np.uint8,
+                            np.uint16, np.uint32, np.uint64)):
+            return int(obj)
+        elif isinstance(obj, (np.float_, np.float16, np.float32,
+                              np.float64)):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        return json.JSONEncoder.default(self, obj)
